requirements.txt
pandas==2.1.0
requests==2.31.0
tqdm==4.66.1
plotly==5.18.0
numpy==1.24.3
python-dotenv==1.0.0
PyGithub==2.1.1